pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yaml-schema-validator"
version = "0.1.0"
description = "A YAML schema validator for GroundX prompt/config files"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.5.0",
    "ruamel.yaml>=0.18.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
yaml-validate = "main:main"

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
where = ["."]
include = ["yaml_validator*"]

[tool.setuptools.package-data]
yaml_validator = ["profiles/configs/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
pythonpath = ["."]
